implement calling functions with arguments
Right now, all arguments are simply passed on the stack in reverse
order, which keeps us from having to implement register allocation yet.
It also doesn't check that correct number of arguments are passed to the
called function - this is impossible in the general case anyway, since B
has no types.